requirements.txt
scipy==1.9.1
pandas==1.5.2
requests==2.28.1
websocket-client==1.5.0
pymongo==4.6.3
python-dotenv==0.20.0
requests-html==0.10.0
python-telegram-bot==13.7
requests-cache==0.9.
kafka-python==2.0.2
pydantic==2.7.0
# pyspark==3.5.1
# pyarrow==15.0.2
# py4j
httpx
confluent-kafka
